adjusted location and resolution of health bar, capped health
Health bar was not showing up on my screen, so I put it so that it is always snapped to the top left corner. Try resizing the game window to see.

I also fixed the filtering mode and resolution of the bar sprite.

Also capped off health to make testing easier