Progress towards player inventory
Ditching local progress on seb-ui branch and continuing on this branch instead since I forgot to make a PR to add the chest to master and I wanted to work late (woops). Don't worry, this system that I am working on in this branch is miles better than the old one in my local seb-ui branch anyway.