Transferred every change from master and seb-chest-2 in this branch
It might be easier for you to go through the commits in the seb-chest-2 branch to see the progression of the inventory system. This branch had to be made to manually transfer the changes to the SampleScene.unity file without breaking anything in master since we are missing a reliable merging tool for this type of file (coming very soon).