Added simple card UI
This code is enough to have 1 card work. It does not implement having multiple cards yet or being able to quickly create cards like I did with the items yet. However, I feel pretty confident about being able to reuse some of the thoughts processes that are present in the items creation and the bag button for that.