Altars send you to the next randomly generated dungeon
Altars are the cylinders. They spawn in a leaf room that the player doesnt spawn in. press E to interact when in range. it increases the dungeon score obtained when dying by 1 each time. Regenerates a new dungeon with new bandits and rooms. Relocates the player to the new spawn point of the new dungeon without having to reinstantiate the core prefab, which allows for the player's current stats/loot/cards to be kept when transfering to the next floor.