Fixed all (or moat) bugs related to the save system.
The only improvements it could have would be the ability to begin from your last saved point in the death menu, and the option to go back to main menu from pause (Unrelated to save). I also began working on trying to fix the gold but didnt do too much.